requirements.txt
pyarrow==13.0.0
datasets==2.14.5
mlflow==2.8.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
fastapi==0.104.0
uvicorn==0.24.0
pyyaml==6.0.1